+++ src/web/Controllers/SettingController.cs
@@ -1,10 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Pika.Lib.Extension;
 using Pika.Lib.Model;
 using Pika.Lib.Store;
 using Pika.Lib.Util;
@@ -15,7 +15,7 @@ public class SettingController : Controller
 {
     private readonly ILogger<SettingController> _logger;
     private readonly IDbRepository _repository;
-    private readonly PikaSetting _setting;
+    private PikaSetting _setting;
 
     public SettingController(PikaSetting setting, IDbRepository repository, ILogger<SettingController> logger)
     {
@@ -26,6 +26,7 @@ public class SettingController : Controller
 
     public IActionResult Index()
     {
+        ViewData["DbSize"] = _repository.GetDbSize().ToByteSizeHuman();
         return View(_setting);
     }
 
@@ -33,9 +34,14 @@ public class SettingController : Controller
     public async Task<IActionResult> ExportAsync()
     {
         var tasks = await _repository.GetTasksAsync(int.MaxValue, 0, orderByClause: "created_at ASC");
+        var export = new PikaExport
+        {
+            Setting = _setting,
+            Tasks = tasks
+        };
         var content =
-            Encoding.UTF8.GetBytes(JsonUtil.Serialize(tasks, true));
-        return File(content, "application/json", "pika-task.json");
+            Encoding.UTF8.GetBytes(JsonUtil.Serialize(export, true));
+        return File(content, "application/json", "pika-export.json");
     }
 
     [HttpPost("/setting/import")]
@@ -44,9 +50,28 @@ public class SettingController : Controller
         try
         {
             await using var stream = file.OpenReadStream();
-            foreach (var pikaTask in await JsonUtil.DeserializeAsync<List<PikaTask>>(stream))
+            var export = await JsonUtil.DeserializeAsync<PikaExport>(stream);
+            if (export.Setting != null)
             {
-                await _repository.AddTaskAsync(pikaTask);
+                if (export.Setting.RetainSizeInMb < 1)
+                {
+                    export.Setting.RetainSizeInMb = _setting.RetainSizeInMb;
+                }
+
+                if (export.Setting.ItemsPerPage < 1 || export.Setting.ItemsPerPage > 50)
+                {
+                    export.Setting.ItemsPerPage = _setting.ItemsPerPage;
+                }
+
+                _setting = export.Setting;
+            }
+
+            if (export.Tasks != null)
+            {
+                foreach (var pikaTask in export.Tasks)
+                {
+                    await _repository.AddTaskAsync(pikaTask);
+                }
             }
 
             return Redirect("~/task");